+++ src/LighthouseServiceProvider.php
@@ -8,11 +8,9 @@ use Illuminate\Contracts\Config\Repository as ConfigRepository;
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Foundation\Application as LaravelApplication;
 use Illuminate\Routing\Router;
-use Illuminate\Support\Arr;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Factory as ValidationFactory;
-use Illuminate\Validation\Validator;
 use Laravel\Lumen\Application as LumenApplication;
 use Nuwave\Lighthouse\Console\ClearCacheCommand;
 use Nuwave\Lighthouse\Console\IdeHelperCommand;
@@ -38,6 +36,7 @@ use Nuwave\Lighthouse\Schema\Source\SchemaSourceProvider;
 use Nuwave\Lighthouse\Schema\Source\SchemaStitcher;
 use Nuwave\Lighthouse\Schema\TypeRegistry;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
+use Nuwave\Lighthouse\Support\AppVersion;
 use Nuwave\Lighthouse\Support\Compatibility\LaravelMiddlewareAdapter;
 use Nuwave\Lighthouse\Support\Compatibility\LumenMiddlewareAdapter;
 use Nuwave\Lighthouse\Support\Compatibility\MiddlewareAdapter;
@@ -51,6 +50,43 @@ use Nuwave\Lighthouse\Support\Http\Responses\ResponseStream;
 
 class LighthouseServiceProvider extends ServiceProvider
 {
+    /**
+     * Bootstrap any application services.
+     *
+     * @param  \Illuminate\Validation\Factory  $validationFactory
+     * @param  \Illuminate\Contracts\Config\Repository  $configRepository
+     * @return void
+     */
+    public function boot(ValidationFactory $validationFactory, ConfigRepository $configRepository): void
+    {
+        $this->publishes([
+            __DIR__.'/lighthouse.php' => $this->app->make('path.config').'/lighthouse.php',
+        ], 'config');
+
+        $this->publishes([
+            __DIR__.'/../assets/default-schema.graphql' => $configRepository->get('lighthouse.schema.register'),
+        ], 'schema');
+
+        $this->loadRoutesFrom(__DIR__.'/Support/Http/routes.php');
+    }
+
+    /**
+     * Load routes from provided path.
+     *
+     * @param  string  $path
+     * @return void
+     */
+    protected function loadRoutesFrom($path): void
+    {
+        if (AppVersion::isLumen()) {
+            require realpath($path);
+
+            return;
+        }
+
+        parent::loadRoutesFrom($path);
+    }
+
     /**
      * Register any application services.
      *
@@ -133,42 +169,5 @@ class LighthouseServiceProvider extends ServiceProvider
                 ValidateSchemaCommand::class,
             ]);
         }
-    }
-
-    /**
-     * Bootstrap any application services.
-     *
-     * @param  \Illuminate\Validation\Factory  $validationFactory
-     * @param  \Illuminate\Contracts\Config\Repository  $configRepository
-     * @return void
-     */
-    public function boot(ValidationFactory $validationFactory, ConfigRepository $configRepository): void
-    {
-        $this->publishes([
-            __DIR__.'/lighthouse.php' => $this->app->make('path.config').'/lighthouse.php',
-        ], 'config');
-
-        $this->publishes([
-            __DIR__.'/../assets/default-schema.graphql' => $configRepository->get('lighthouse.schema.register'),
-        ], 'schema');
-
-        $this->loadRoutesFrom(__DIR__.'/Support/Http/routes.php');
-    }
-
-    /**
-     * Load routes from provided path.
-     *
-     * @param  string  $path
-     * @return void
-     */
-    protected function loadRoutesFrom($path): void
-    {
-        if (Str::contains($this->app->version(), 'Lumen')) {
-            require realpath($path);
-
-            return;
-        }
-
-        parent::loadRoutesFrom($path);
     }
 }