+++ src/Testing/MakesGraphQLRequests.php
@@ -3,7 +3,6 @@
 namespace Nuwave\Lighthouse\Testing;
 
 use GraphQL\Type\Introspection;
-use Illuminate\Foundation\Testing\TestResponse;
 use Illuminate\Support\Arr;
 use Nuwave\Lighthouse\Support\Contracts\CanStreamResponse;
 use Nuwave\Lighthouse\Support\Http\Responses\MemoryStream;
@@ -23,7 +22,7 @@ trait MakesGraphQLRequests
      * On the first call to introspect() this property is set to
      * cache the result, as introspection is quite expensive.
      *
-     * @var \Illuminate\Foundation\Testing\TestResponse|null
+     * @var \Illuminate\Foundation\Testing\TestResponse|\Illuminate\Testing\TestResponse|null
      */
     protected $introspectionResult;
 
@@ -40,9 +39,9 @@ trait MakesGraphQLRequests
      * @param  string  $query
      * @param  array|null  $variables
      * @param  array  $extraParams
-     * @return \Illuminate\Foundation\Testing\TestResponse
+     * @return \Illuminate\Foundation\Testing\TestResponse|\Illuminate\Testing\TestResponse
      */
-    protected function graphQL(string $query, array $variables = null, array $extraParams = []): TestResponse
+    protected function graphQL(string $query, array $variables = null, array $extraParams = [])
     {
         $params = ['query' => $query];
 
@@ -60,9 +59,9 @@ trait MakesGraphQLRequests
      *
      * @param  mixed[]  $data
      * @param  mixed[]  $headers
-     * @return \Illuminate\Foundation\Testing\TestResponse
+     * @return \Illuminate\Foundation\Testing\TestResponse|\Illuminate\Testing\TestResponse
      */
-    protected function postGraphQL(array $data, array $headers = []): TestResponse
+    protected function postGraphQL(array $data, array $headers = [])
     {
         return $this->postJson(
             $this->graphQLEndpointUrl(),
@@ -79,9 +78,9 @@ trait MakesGraphQLRequests
      *
      * @param  mixed[]  $parameters
      * @param  mixed[]  $files
-     * @return \Illuminate\Foundation\Testing\TestResponse
+     * @return \Illuminate\Foundation\Testing\TestResponse|\Illuminate\Testing\TestResponse
      */
-    protected function multipartGraphQL(array $parameters, array $files): TestResponse
+    protected function multipartGraphQL(array $parameters, array $files)
     {
         return $this->call(
             'POST',
@@ -98,9 +97,9 @@ trait MakesGraphQLRequests
     /**
      * Execute the introspection query on the GraphQL server.
      *
-     * @return \Illuminate\Foundation\Testing\TestResponse
+     * @return \Illuminate\Foundation\Testing\TestResponse|\Illuminate\Testing\TestResponse
      */
-    protected function introspect(): TestResponse
+    protected function introspect()
     {
         if ($this->introspectionResult) {
             return $this->introspectionResult;
@@ -152,7 +151,7 @@ trait MakesGraphQLRequests
 
         return Arr::first(
             $results,
-            function (array $result) use ($name): bool {
+            static function (array $result) use ($name): bool {
                 return $result['name'] === $name;
             }
         );