+++ tests/Integration/Schema/Directives/CanDirectiveDBTest.php
@@ -4,6 +4,7 @@ namespace Tests\Integration\Schema\Directives;
 
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Nuwave\Lighthouse\Exceptions\AuthorizationException;
+use Nuwave\Lighthouse\Schema\Directives\CanDirective;
 use Tests\DBTestCase;
 use Tests\Utils\Models\Post;
 use Tests\Utils\Models\Task;
@@ -63,8 +64,7 @@ class CanDirectiveDBTest extends DBTestCase
             $this->never()
         );
 
-        $this->schema = /** @lang GraphQL */
-            '
+        $this->schema = /** @lang GraphQL */ '
         type Query {
             user(id: ID @eq): User
                 @can(ability: "view", find: "id")
@@ -85,6 +85,86 @@ class CanDirectiveDBTest extends DBTestCase
             }
         }
         ');
+    }
+
+    public function testThrowsIfFindValueIsNotGiven(): void
+    {
+        $this->be(
+            new User([
+                'name' => UserPolicy::ADMIN,
+            ])
+        );
+
+        $this->schema = /** @lang GraphQL */ '
+        type Query {
+            user(id: ID): User
+                @can(ability: "view", find: "some.path")
+                @first
+        }
+
+        type User {
+            id: ID!
+            name: String!
+        }
+        ';
+
+        $this->graphQL(/** @lang GraphQL */ '
+        {
+            user {
+                name
+            }
+        }
+        ')->assertJson([
+            'errors' => [
+                [
+
+                    'message' => CanDirective::missingKeyToFindModel('some.path'),
+                ],
+            ],
+        ]);
+    }
+
+    public function testFindUsingNestedInputWithDotNotation(): void
+    {
+        $user = factory(User::class)->create([
+            'name' => 'foo',
+        ]);
+        $this->be($user);
+
+        $this->schema = /** @lang GraphQL */ '
+        type Query {
+            user(input: FindUserInput): User
+                @can(ability: "view", find: "input.id")
+                @first
+        }
+
+        type User {
+            id: ID!
+            name: String!
+        }
+
+        input FindUserInput {
+          id: ID
+        }
+        ';
+
+        $this->graphQL(/** @lang GraphQL */ '
+        query ($id: ID){
+            user(input: {
+              id: $id
+            }) {
+                name
+            }
+        }
+        ', [
+            'id' => $user->id,
+        ])->assertJson([
+            'data' => [
+                'user' => [
+                    'name' => 'foo',
+                ],
+            ],
+        ]);
     }
 
     public function testThrowsIfNotAuthorized(): void